feat: ActionMessage refactor v1 **wip**
