+++ apps/marginfi-v2-ui/src/components/common/Portfolio/components/move-position/index.ts
@@ -1,0 +1,1 @@
+export * from "./move-position-dialog";
