+++ apps/marginfi-v2-ui/src/components/common/Portfolio/hooks/index.ts
@@ -1,1 +1,2 @@
 export * from "./use-reward-simulation";
+export * from "./use-move-position-simulation";
