+++ packages/mrgn-ui/src/components/action-box-v2/actions/lend-box/utils/index.ts
@@ -1,3 +1,2 @@
-export * from "./lend-box.utils";
 export * from "./lend-simulation.utils";
 export * from "./lend-actions.utils";
