--- packages/mrgn-ui/src/components/action-box-v2/actions/lend-box/utils/lend-box.utils.ts
@@ -1,103 +1,0 @@
-import {
-  ActionMethod,
-  ActionMethodType,
-  canBeBorrowed,
-  canBeLent,
-  canBeRepaid,
-  canBeWithdrawn,
-} from "@mrgnlabs/mrgn-utils";
-import { ActionType, ExtendedBankInfo } from "@mrgnlabs/marginfi-v2-ui-state";
-import { MarginfiAccountWrapper } from "@mrgnlabs/marginfi-client-v2";
-
-export function getColorForActionMethodType(type?: ActionMethodType) {
-  if (type === "INFO") {
-    return "info";
-  } else if (type === "WARNING") {
-    return "alert";
-  } else {
-    return "alert";
-  }
-}
-
-interface CheckActionAvailableProps {
-  amount: number | null;
-  connected: boolean;
-  nativeSolBalance: number;
-  showCloseBalance?: boolean;
-  selectedBank: ExtendedBankInfo | null;
-  banks: ExtendedBankInfo[];
-  marginfiAccount: MarginfiAccountWrapper | null;
-  lendMode: ActionType;
-}
-
-export function checkActionAvailable({
-  amount,
-  nativeSolBalance,
-  connected,
-  showCloseBalance,
-  selectedBank,
-  banks,
-  marginfiAccount,
-  lendMode,
-}: CheckActionAvailableProps): ActionMethod[] {
-  let checks: ActionMethod[] = [];
-
-  const requiredCheck = getRequiredCheck(connected, selectedBank);
-  if (requiredCheck) return [requiredCheck];
-
-  const generalChecks = getGeneralChecks(amount ?? 0, showCloseBalance);
-  if (generalChecks) checks.push(...generalChecks);
-
-  // allert checks
-  if (selectedBank) {
-    switch (lendMode) {
-      case ActionType.Deposit:
-        const lentChecks = canBeLent(selectedBank, nativeSolBalance);
-        if (lentChecks.length) checks.push(...lentChecks);
-        break;
-      case ActionType.Withdraw:
-        const withdrawChecks = canBeWithdrawn(selectedBank, marginfiAccount);
-        if (withdrawChecks.length) checks.push(...withdrawChecks);
-        break;
-      case ActionType.Borrow:
-        const borrowChecks = canBeBorrowed(selectedBank, banks, marginfiAccount);
-        if (borrowChecks.length) checks.push(...borrowChecks);
-        break;
-      case ActionType.Repay:
-        const repayChecks = canBeRepaid(selectedBank);
-        if (repayChecks) checks.push(...repayChecks);
-        break;
-    }
-  }
-
-  if (checks.length === 0)
-    checks.push({
-      isEnabled: true,
-    });
-
-  return checks;
-}
-
-function getRequiredCheck(connected: boolean, selectedBank: ExtendedBankInfo | null): ActionMethod | null {
-  if (!connected) {
-    return { isEnabled: false };
-  }
-  if (!selectedBank) {
-    return { isEnabled: false };
-  }
-
-  return null;
-}
-
-function getGeneralChecks(amount: number = 0, showCloseBalance?: boolean): ActionMethod[] {
-  let checks: ActionMethod[] = [];
-  if (showCloseBalance) {
-    checks.push({ actionMethod: "INFO", description: "Close lending balance.", isEnabled: true });
-  }
-
-  if (amount === 0) {
-    checks.push({ isEnabled: false });
-  }
-
-  return checks;
-}
