+++ packages/mrgn-ui/src/components/action-box-v2/actions/loop-box/utils/index.ts
@@ -1,3 +1,2 @@
 export * from "./loop-simulation.utils";
 export * from "./loop-action.utils";
-export * from "./loop-box.utils";
