+++ packages/mrgn-ui/src/components/action-box-v2/actions/repay-collat-box/repay-collat-box.tsx
@@ -8,15 +8,22 @@ import {
   DEFAULT_ACCOUNT_SUMMARY,
   ActiveBankInfo,
 } from "@mrgnlabs/marginfi-v2-ui-state";
-import { ActionMethod, MarginfiActionParams, PreviousTxn, showErrorToast } from "@mrgnlabs/mrgn-utils";
+import {
+  ActionMethod,
+  checkRepayCollatActionAvailable,
+  MarginfiActionParams,
+  PreviousTxn,
+  showErrorToast,
+} from "@mrgnlabs/mrgn-utils";
 import { MarginfiAccountWrapper, MarginfiClient } from "@mrgnlabs/marginfi-client-v2";
 
 import { CircularProgress } from "~/components/ui/circular-progress";
 import { Tooltip, TooltipContent, TooltipProvider, TooltipTrigger } from "~/components/ui/tooltip";
-import { ActionButton, ActionMessage, ActionSettingsButton } from "~/components/action-box-v2/components";
+import { ActionButton, ActionSettingsButton } from "~/components/action-box-v2/components";
 import { useActionAmounts, usePollBlockHeight } from "~/components/action-box-v2/hooks";
+import { ActionMessage } from "~/components";
 
-import { checkActionAvailable, handleExecuteRepayCollatAction } from "./utils";
+import { handleExecuteRepayCollatAction } from "./utils";
 import { Collateral, ActionInput, Preview } from "./components";
 import { useRepayCollatBoxStore } from "./store";
 import { useRepayCollatSimulation } from "./hooks";
@@ -172,7 +179,7 @@ export const RepayCollatBox = ({
 
   const actionMethods = React.useMemo(
     () =>
-      checkActionAvailable({
+      checkRepayCollatActionAvailable({
         amount,
         connected,
         selectedBank,
@@ -324,3 +331,12 @@ export const RepayCollatBox = ({
     </>
   );
 };
+function checkRepayColatActionAvailable(arg0: {
+  amount: number;
+  connected: boolean;
+  selectedBank: ExtendedBankInfo | null;
+  selectedSecondaryBank: ExtendedBankInfo | null;
+  actionQuote: import("@jup-ag/api").QuoteResponse | null;
+}): any {
+  throw new Error("Function not implemented.");
+}
