+++ packages/mrgn-ui/src/components/action-box-v2/actions/repay-collat-box/utils/index.ts
@@ -1,3 +1,2 @@
 export * from "./repay-simulation.utils";
 export * from "./repay-action.utils";
-export * from "./repay-box.utils";
