+++ packages/mrgn-ui/src/components/action-box-v2/actions/stake-box/stake-box.tsx
@@ -2,23 +2,23 @@ import React, { useEffect } from "react";
 
 import { WalletContextState } from "@solana/wallet-adapter-react";
 
-import { getPriceWithConfidence, MarginfiAccountWrapper, MarginfiClient } from "@mrgnlabs/marginfi-client-v2";
+import { getPriceWithConfidence, MarginfiClient } from "@mrgnlabs/marginfi-client-v2";
 import { AccountSummary, ActionType, ExtendedBankInfo } from "@mrgnlabs/marginfi-v2-ui-state";
-import { ActionMethod, LstData, PreviousTxn, showErrorToast, STATIC_SIMULATION_ERRORS } from "@mrgnlabs/mrgn-utils";
+import { ActionMethod, checkStakeActionAvailable, LstData, PreviousTxn } from "@mrgnlabs/mrgn-utils";
 import { nativeToUi, NATIVE_MINT as SOL_MINT, uiToNative } from "@mrgnlabs/mrgn-common";
 
-import { useActionAmounts, usePollBlockHeight } from "~/components/action-box-v2/hooks";
+import { useActionAmounts } from "~/components/action-box-v2/hooks";
 import { WalletContextStateOverride } from "~/components/wallet-v2/hooks/use-wallet.hook";
+import { ActionMessage } from "~/components";
 
 import { useStakeBoxStore } from "./store";
 import { AmountPreview } from "./components/amount-preview";
-import { ActionButton, ActionMessage, ActionSettingsButton } from "../../components";
+import { ActionButton, ActionSettingsButton } from "../../components";
 import { StatsPreview } from "./components/stats-preview";
 import { useStakeSimulation } from "./hooks";
 import { useActionBoxStore } from "../../store";
 import { handleExecuteLstAction } from "./utils/stake-action.utils";
 import { ActionInput } from "./components/action-input";
-import { checkActionAvailable } from "./utils";
 import { useStakeBoxContext } from "../../contexts";
 
 export type StakeBoxProps = {
@@ -236,7 +236,7 @@ export const StakeBox = ({
 
   const actionMethods = React.useMemo(() => {
     setAdditionalActionMethods([]);
-    return checkActionAvailable({
+    return checkStakeActionAvailable({
       amount,
       connected,
       selectedBank,
