+++ packages/mrgn-ui/src/components/action-box-v2/actions/stake-box/utils/index.ts
@@ -1,2 +1,1 @@
 export * from "./stake-simulation.utils";
-export * from "./stake-box.utils";
