+++ packages/mrgn-ui/src/components/index.ts
@@ -2,3 +2,4 @@ export * from "./action-box-v2";
 export * from "./wallet-v2";
 export * from "./LSTDialog";
 export * from "./action-complete";
+export * from "./action-message";
